# docs/swagger.json
{
    "schemes": [
        "HTTP"
    ],
    "swagger": "2.0",
    "info": {
        "description": "下一代媒体刮削\u0026整理工具",
        "title": "MediaTools API 文档",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/library/archive": {
            "post": {
                "description": "整理一个视频文件及其相关的字幕和音轨文件到指定目录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "归档媒体文件",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.LibraryArchiveMediaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/schemas.FileInfo"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/scrape/video": {
            "post": {
                "description": "刮削视频文件的元数据和相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape"
                ],
                "summary": "刮削视频",
                "parameters": [
                    {
                        "description": "刮削请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.ScrapeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刮削成功",
                        "schema": {
                            "$ref": "#/definitions/schemas.FileInfo"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "刮削失败",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "meta.MediaType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "MediaTypeMovie": "电影",
                "MediaTypeTV": "电视剧",
                "MediaTypeUnknown": "未知"
            },
            "x-enum-descriptions": [
                "未知",
                "电影",
                "电视剧"
            ],
            "x-enum-varnames": [
                "MediaTypeUnknown",
                "MediaTypeMovie",
                "MediaTypeTV"
            ]
        },
...
        "schemas.ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "schemas.FileInfo": {
            "type": "object",
            "properties": {
                "ext": {
                    "description": "文件扩展名",
                    "type": "string"
                },
                "is_dir": {
                    "description": "是否为目录",
                    "type": "boolean"
                },
                "mod_time": {
                    "description": "文件修改时间",
                    "type": "string"
                },
                "name": {
                    "description": "文件名",
                    "type": "string"
                },
                "path": {
                    "description": "文件路径",
                    "type": "string"
                },
                "size": {
                    "description": "文件大小",
                    "type": "integer"
                },
                "storage_type": {
                    "description": "存储系统类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.StorageType"
                        }
                    ]
                }
            }
        },
        "schemas.FileInfoRequest": {
            "type": "object",
            "required": [
                "path",
                "storage_type"
            ],
            "properties": {
                "path": {
                    "type": "string"
                },
                "storage_type": {
                    "$ref": "#/definitions/schemas.StorageType"
                }
            }
        },
        "schemas.LibraryArchiveMediaRequest": {
            "type": "object",
            "required": [
                "dst_dir",
                "src_file"
            ],
            "properties": {
                "dst_dir": {
                    "$ref": "#/definitions/schemas.FileInfoRequest"
                },
                "need_scrape": {
                    "type": "boolean"
                },
                "src_file": {
                    "$ref": "#/definitions/schemas.FileInfoRequest"
                },
                "transfer_type": {
                    "$ref": "#/definitions/schemas.TransferType"
                }
            }
        },
...
        "schemas.ScrapeRequest": {
            "type": "object",
            "required": [
                "dst_file"
            ],
            "properties": {
                "dst_file": {
                    "$ref": "#/definitions/schemas.FileInfoRequest"
                },
                "media_type": {
                    "$ref": "#/definitions/meta.MediaType"
                },
                "tmdb_id": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.StorageType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1
            ],
            "x-enum-comments": {
                "StorageLocal": "本地文件系统",
                "StorageUnknown": "未知文件系统"
            },
            "x-enum-descriptions": [
                "未知文件系统",
                "本地文件系统"
            ],
            "x-enum-varnames": [
                "StorageUnknown",
                "StorageLocal"
            ]
        },
        "schemas.TransferType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-comments": {
                "TransferCopy": "复制",
                "TransferLink": "硬链接",
                "TransferMove": "移动",
                "TransferSoftLink": "软链接",
                "TransferUnknown": "未知传输类型"
            },
            "x-enum-descriptions": [
                "未知传输类型",
                "复制",
                "移动",
                "硬链接",
                "软链接"
            ],
            "x-enum-varnames": [
                "TransferUnknown",
                "TransferCopy",
                "TransferMove",
                "TransferLink",
                "TransferSoftLink"
            ]
        }
    }
}